CODE RUB : BuyerPostsRepository And TimeConstant and BuyerPostEnum An…

+++ src/GreenSale.Service/Service/BuyerPosts/BuyerPostService.cs
@@ -21,6 +21,7 @@ public class BuyerPostService : IBuyerPostService
     private readonly IFileService _fileService;
     private readonly IBuyerPostImageRepository _imageRepository;
     private readonly IIdentityService _identity;
+    private readonly string BUYERPOSTIMAGES = "BuyerPostImages";
 
     public BuyerPostService(
         IBuyerPostRepository postRepository,
@@ -70,12 +71,14 @@ public class BuyerPostService : IBuyerPostService
         {
             foreach (var item in dto.ImagePath)
             {
-                var img = await _fileService.UploadImageAsync(item);
+                var img = await _fileService.UploadImageAsync(item, BUYERPOSTIMAGES);
 
                 BuyerPostImage BuyerPostImage = new BuyerPostImage()
                 {
                     BuyerpostId = DbResult,
                     ImagePath = img,
+                    CreatedAt = TimeHelper.GetDateTime(),
+                    UpdatedAt = TimeHelper.GetDateTime(),
                 };
 
                 var DbImgResult = await _imageRepository.CreateAsync(BuyerPostImage);
@@ -94,7 +97,21 @@ public class BuyerPostService : IBuyerPostService
         if (DbFound.Id == 0)
             throw new BuyerPostNotFoundException();
 
+        var DbImgAll = await _imageRepository.GetByIdAllAsync(buyerId);
+
+        if(DbImgAll.Count == 0)
+            throw new BuyerPostNotFoundException();
+
+        var DbImgResult = await _imageRepository.DeleteAsync(buyerId);
         var DbResult = await _postRepository.DeleteAsync(buyerId);
+
+        if (DbResult > 0 && 0 < DbImgResult)
+        {
+            foreach (var item in DbImgAll)
+            {
+                await _fileService.DeleteImageAsync(item.ImagePath);
+            }
+        }
 
         return DbResult > 0;
     }
@@ -102,36 +119,122 @@ public class BuyerPostService : IBuyerPostService
     public async Task<List<BuyerPostViewModel>> GetAllAsync(PaginationParams @params)
     {
         var DbResult = await _postRepository.GetAllAsync(@params);
+        var dBim = await _imageRepository.GetAllAsync(@params);
+        List<BuyerPostViewModel> Result = new List<BuyerPostViewModel>();
+        BuyerPostViewModel buyerPostViewModel = new BuyerPostViewModel();
+
+        foreach (var item in DbResult)
+        {
+            buyerPostViewModel = new BuyerPostViewModel()
+            {
+                Id = item.Id,
+                FullName = item.FullName,
+                CategoryName = item.CategoryName,
+                Title = item.Title,
+                Description = item.Description,
+                Price = item.Price,
+                Capacity = item.Capacity,
+                CapacityMeasure = item.CapacityMeasure,
+                Type = item.Type,
+                Region = item.Region,
+                District = item.District,
+                Address = item.Address,
+                Status = item.Status,
+                CreatedAt = item.CreatedAt,
+                UpdatedAt = item.UpdatedAt,
+            };
+            buyerPostViewModel.BuyerPostsImages = new List<BuyerPostImage>();
+            foreach (var img in dBim)
+            {
+                if(img.BuyerpostId == item.Id)
+                {
+                    
+                    BuyerPostImage buyerPostImage = new BuyerPostImage()
+                    {
+                        Id = img.Id,
+                        BuyerpostId = img.BuyerpostId,
+                        ImagePath = img.ImagePath,
+                        CreatedAt = img.CreatedAt,
+                        UpdatedAt = img.UpdatedAt,
+                    };
+
+                    buyerPostViewModel.BuyerPostsImages.Add(buyerPostImage);
+                }
+            }
+            Result.Add(buyerPostViewModel);
+        }
+
         var DBCount = await _postRepository.CountAsync();
         _paginator.Paginate(DBCount, @params);
 
-        return DbResult;
+        return Result;
     }
 
     public async Task<BuyerPostViewModel> GetBYIdAsync(long buyerId)
     {
-        var DbFound = await _postRepository.GetByIdAsync(buyerId);
+        var item = await _postRepository.GetByIdAsync(buyerId);
+        var dBim = await _imageRepository.GetByIdAllAsync(buyerId);
 
-        if (DbFound.Id == 0)
+        if (item.Id == 0)
             throw new BuyerPostNotFoundException();
 
-        return DbFound;
+        BuyerPostViewModel buyerPostViewModel = new BuyerPostViewModel();
+
+        buyerPostViewModel = new BuyerPostViewModel()
+        {
+            Id = item.Id,
+            FullName = item.FullName,
+            CategoryName = item.CategoryName,
+            Title = item.Title,
+            Description = item.Description,
+            Price = item.Price,
+            Capacity = item.Capacity,
+            CapacityMeasure = item.CapacityMeasure,
+            Type = item.Type,
+            Region = item.Region,
+            District = item.District,
+            Address = item.Address,
+            Status = item.Status,
+            CreatedAt = item.CreatedAt,
+            UpdatedAt = item.UpdatedAt,
+        };
+        buyerPostViewModel.BuyerPostsImages = new List<BuyerPostImage>();
+        foreach (var img in dBim)
+        {
+            if (img.BuyerpostId == item.Id)
+            {
+
+                BuyerPostImage buyerPostImage = new BuyerPostImage()
+                {
+                    Id = img.Id,
+                    BuyerpostId = img.BuyerpostId,
+                    ImagePath = img.ImagePath,
+                    CreatedAt = img.CreatedAt,
+                    UpdatedAt = img.UpdatedAt,
+                };
+
+                buyerPostViewModel.BuyerPostsImages.Add(buyerPostImage);
+            }
+        }
+
+        return buyerPostViewModel;
     }
 
     public async Task<bool> ImageUpdateAsync(BuyerPostImageDto dto)
     {
         var DbFoundImg = await _imageRepository.GetByIdAsync(dto.BuyerPostImageId);
 
-        if (DbFoundImg is null)
+        if (DbFoundImg.Id == 0)
             throw new ImageNotFoundException();
 
         var RootDEl = await _fileService.DeleteImageAsync(DbFoundImg.ImagePath);
-        var img = await _fileService.UploadImageAsync(dto.ImagePath);
+        var img = await _fileService.UploadImageAsync(dto.ImagePath, BUYERPOSTIMAGES);
 
         BuyerPostImage buyerPostImage = new BuyerPostImage()
         {
             BuyerpostId = dto.BuyerPostId,
             ImagePath = img,
+            UpdatedAt = TimeHelper.GetDateTime(),
         };
 
         var DbResult = await _imageRepository.UpdateAsync(dto.BuyerPostImageId, buyerPostImage);
@@ -160,7 +263,7 @@ public class BuyerPostService : IBuyerPostService
             Address = dto.Address,
             District = dto.District,
             PhoneNumber = dto.PhoneNumber,
-            Status = Domain.Enums.BuyerPosts.BuyerPostEnum.New,
+            Status = dto.Status,
             UpdatedAt = TimeHelper.GetDateTime(),
         };
